Update WordPress Settings plugin metadata and remove extra lines
The changes in this commit include renaming the name of the WordPress Settings plugin to 'WordPress Settings Wrapper' and altering its description to 'WordPress Settings Library as Plugin'. The version was also changed to 'develop'. Additionally, unnecessary new lines were removed to improve the code cleanliness in both the main plugin file and the test controller file.