Add Github actions workflow for plugin build and release
This commit introduces a continuous integration (CI) workflow for PHPCS and ESLint checks. It further covers the build and release steps, generating a zip file for each new version of the WordPress plugin. The build is triggered on push to main or on new tags, while the release part is only executed on new tags.