Remove gitlab-ci.yml and update package file and README
The `.gitlab-ci.yml` file in the Wordpress settings plugin has been deleted. It no longer needed for the project build. Additionally, the package.json file was updated with a minor change where the `js` directory is removed from the files array. Lastly, the installation instructions have been expanded in the README file with the composer require command.